Seperated SocketIO Connection
