+++ .gitignore
@@ -1,5 +1,6 @@
 
 node_modules
+build
 
 
 # Dont push env variables to github
