+++ package.json
@@ -11,7 +11,8 @@
     "Rohit Prajapati <roht.praz@gmail.com>"
   ],
   "scripts": {
-    "start": "nodemon src/index.js",
+    "start": "nodemon src/index.ts",
+    "build": "tsc --project ./",
     "test": "jest"
   },
   "license": "MIT",
@@ -20,12 +21,18 @@
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
     "nodemon": "^2.0.7",
-    "socket.io": "^4.1.2"
+    "socket.io": "^4.1.2",
+    "socket.io-client": "^4.1.2"
   },
   "devDependencies": {
     "@types/express": "^4.17.12",
     "@types/jest": "^26.0.23",
+    "@types/node": "^15.12.2",
+    "@types/supertest": "^2.0.11",
     "jest": "^27.0.4",
-    "supertest": "^6.1.3"
+    "supertest": "^6.1.3",
+    "ts-jest": "^27.0.3",
+    "ts-node": "^10.0.0",
+    "typescript": "^4.3.2"
   }
 }
